ex8

+++ C++程序设计课程设计/studentcpp1.cpp
@@ -28,13 +28,13 @@ void student::modify()
 	get_classname();
 	get_abstype();
 }
-int student::check(int i,int d,int o)
+int student::check(long long i,int d,int o)
 {
 	if(id == i && d == date && o == order)
 	return 1;
 	else return 0;
 }
-void student::search(int n)
+void student::search(long long n)
 {
 	if(id == n)
 	{
